refactor(framework/agent): extract dispatch and bridge signature parsing.
Move agent event handling out of NovuRequestHandler into agent-dispatch,
and share novu-signature header parsing via utils/bridge-signature.
